+++ spring-integration-core/src/main/java/org/springframework/integration/aot/CoreRuntimeHints.java
@@ -39,14 +39,15 @@ import org.springframework.context.SmartLifecycle;
 import org.springframework.integration.aggregator.MessageGroupProcessor;
 import org.springframework.integration.context.IntegrationContextUtils;
 import org.springframework.integration.context.IntegrationProperties;
+import org.springframework.integration.core.GenericHandler;
 import org.springframework.integration.core.GenericSelector;
+import org.springframework.integration.core.GenericTransformer;
 import org.springframework.integration.core.MessageSource;
 import org.springframework.integration.dsl.IntegrationFlow;
 import org.springframework.integration.gateway.MethodArgsHolder;
 import org.springframework.integration.gateway.RequestReplyExchanger;
 import org.springframework.integration.handler.AbstractReplyProducingMessageHandler;
 import org.springframework.integration.handler.DelayHandler;
-import org.springframework.integration.handler.GenericHandler;
 import org.springframework.integration.history.MessageHistory;
 import org.springframework.integration.json.JsonPathUtils;
 import org.springframework.integration.message.AdviceMessage;
@@ -56,7 +57,6 @@ import org.springframework.integration.store.MessageHolder;
 import org.springframework.integration.store.MessageMetadata;
 import org.springframework.integration.support.MutableMessage;
 import org.springframework.integration.support.MutableMessageHeaders;
-import org.springframework.integration.transformer.GenericTransformer;
 import org.springframework.messaging.MessageHandler;
 import org.springframework.messaging.MessageHeaders;
 import org.springframework.messaging.PollableChannel;
