+++ spring-integration-core/src/main/java/org/springframework/integration/config/AggregatorAnnotationPostProcessor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.integration.config.annotation;
+package org.springframework.integration.config;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
