+++ spring-integration-core/src/main/java/org/springframework/integration/handler/LambdaMessageProcessor.java
@@ -32,8 +32,9 @@ import org.springframework.beans.factory.BeanFactoryAware;
 import org.springframework.core.MethodIntrospector;
 import org.springframework.core.log.LogMessage;
 import org.springframework.integration.context.IntegrationContextUtils;
+import org.springframework.integration.core.GenericHandler;
 import org.springframework.integration.core.GenericSelector;
-import org.springframework.integration.transformer.GenericTransformer;
+import org.springframework.integration.core.GenericTransformer;
 import org.springframework.lang.Nullable;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.MessageHeaders;
