+++ spring-integration-core/src/main/java/org/springframework/integration/transformer/Transformer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2019 the original author or authors.
+ * Copyright 2002-2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,6 +16,7 @@
 
 package org.springframework.integration.transformer;
 
+import org.springframework.integration.core.GenericTransformer;
 import org.springframework.messaging.Message;
 
 /**
