+++ spring-integration-core/src/test/java/org/springframework/integration/config/annotation/SubscriberOrderTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2019 the original author or authors.
+ * Copyright 2002-2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,7 +21,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import org.springframework.beans.factory.support.BeanDefinitionBuilder;
 import org.springframework.beans.factory.support.RootBeanDefinition;
@@ -31,6 +31,7 @@ import org.springframework.integration.MessageRejectedException;
 import org.springframework.integration.annotation.MessageEndpoint;
 import org.springframework.integration.annotation.ServiceActivator;
 import org.springframework.integration.channel.DirectChannel;
+import org.springframework.integration.config.MessagingAnnotationPostProcessor;
 import org.springframework.integration.dispatcher.RoundRobinLoadBalancingStrategy;
 import org.springframework.integration.test.util.TestUtils;
 import org.springframework.messaging.Message;
@@ -45,99 +46,103 @@ public class SubscriberOrderTests {
 
 	@Test
 	public void directChannelAndFailoverDispatcherWithSingleCallPerMethod() {
-		GenericApplicationContext context = TestUtils.createTestApplicationContext();
-		context.registerBeanDefinition("postProcessor", new RootBeanDefinition(MessagingAnnotationPostProcessor.class));
-		RootBeanDefinition channelDefinition = new RootBeanDefinition(DirectChannel.class);
-		context.registerBeanDefinition("input", channelDefinition);
-		RootBeanDefinition testBeanDefinition = new RootBeanDefinition(TestBean.class);
-		testBeanDefinition.getConstructorArgumentValues().addGenericArgumentValue(1);
-		context.registerBeanDefinition("testBean", testBeanDefinition);
-		context.refresh();
-		TestBean testBean = (TestBean) context.getBean("testBean");
-		MessageChannel channel = (MessageChannel) context.getBean("input");
-		channel.send(new GenericMessage<String>("test-1"));
-		channel.send(new GenericMessage<String>("test-2"));
-		channel.send(new GenericMessage<String>("test-3"));
-		channel.send(new GenericMessage<String>("test-4"));
-		channel.send(new GenericMessage<String>("test-5"));
-		List<Integer> calls = testBean.calls;
-		assertThat(calls.size()).isEqualTo(5);
-		assertThat(calls.get(0).intValue()).isEqualTo(1);
-		assertThat(calls.get(1).intValue()).isEqualTo(2);
-		assertThat(calls.get(2).intValue()).isEqualTo(3);
-		assertThat(calls.get(3).intValue()).isEqualTo(4);
-		assertThat(calls.get(4).intValue()).isEqualTo(5);
-		context.close();
+		try (GenericApplicationContext context = TestUtils.createTestApplicationContext()) {
+			context.registerBeanDefinition("postProcessor",
+					new RootBeanDefinition(MessagingAnnotationPostProcessor.class));
+			RootBeanDefinition channelDefinition = new RootBeanDefinition(DirectChannel.class);
+			context.registerBeanDefinition("input", channelDefinition);
+			RootBeanDefinition testBeanDefinition = new RootBeanDefinition(TestBean.class);
+			testBeanDefinition.getConstructorArgumentValues().addGenericArgumentValue(1);
+			context.registerBeanDefinition("testBean", testBeanDefinition);
+			context.refresh();
+			TestBean testBean = (TestBean) context.getBean("testBean");
+			MessageChannel channel = (MessageChannel) context.getBean("input");
+			channel.send(new GenericMessage<>("test-1"));
+			channel.send(new GenericMessage<>("test-2"));
+			channel.send(new GenericMessage<>("test-3"));
+			channel.send(new GenericMessage<>("test-4"));
+			channel.send(new GenericMessage<>("test-5"));
+			List<Integer> calls = testBean.calls;
+			assertThat(calls.size()).isEqualTo(5);
+			assertThat(calls.get(0).intValue()).isEqualTo(1);
+			assertThat(calls.get(1).intValue()).isEqualTo(2);
+			assertThat(calls.get(2).intValue()).isEqualTo(3);
+			assertThat(calls.get(3).intValue()).isEqualTo(4);
+			assertThat(calls.get(4).intValue()).isEqualTo(5);
+		}
 	}
 
 	@Test
 	public void directChannelAndFailoverDispatcherWithMultipleCallsPerMethod() {
-		GenericApplicationContext context = TestUtils.createTestApplicationContext();
-		context.registerBeanDefinition("postProcessor", new RootBeanDefinition(MessagingAnnotationPostProcessor.class));
-		BeanDefinitionBuilder channelBuilder = BeanDefinitionBuilder.rootBeanDefinition(DirectChannel.class);
-		channelBuilder.addConstructorArgValue(null);
-		RootBeanDefinition channelDefinition = (RootBeanDefinition) channelBuilder.getBeanDefinition();
-		context.registerBeanDefinition("input", channelDefinition);
-		RootBeanDefinition testBeanDefinition = new RootBeanDefinition(TestBean.class);
-		testBeanDefinition.getConstructorArgumentValues().addGenericArgumentValue(2);
-		context.registerBeanDefinition("testBean", testBeanDefinition);
-		context.refresh();
-		TestBean testBean = (TestBean) context.getBean("testBean");
-		MessageChannel channel = (MessageChannel) context.getBean("input");
-		channel.send(new GenericMessage<String>("test-1"));
-		channel.send(new GenericMessage<String>("test-2"));
-		channel.send(new GenericMessage<String>("test-3"));
-		channel.send(new GenericMessage<String>("test-4"));
-		channel.send(new GenericMessage<String>("test-5"));
-		channel.send(new GenericMessage<String>("test-6"));
-		channel.send(new GenericMessage<String>("test-7"));
-		channel.send(new GenericMessage<String>("test-8"));
-		channel.send(new GenericMessage<String>("test-9"));
-		channel.send(new GenericMessage<String>("test-10"));
-		assertThat(testBean.calls.size()).isEqualTo(10);
-		assertThat(testBean.calls.get(0).intValue()).isEqualTo(1);
-		assertThat(testBean.calls.get(1).intValue()).isEqualTo(1);
-		assertThat(testBean.calls.get(2).intValue()).isEqualTo(2);
-		assertThat(testBean.calls.get(3).intValue()).isEqualTo(2);
-		assertThat(testBean.calls.get(4).intValue()).isEqualTo(3);
-		assertThat(testBean.calls.get(5).intValue()).isEqualTo(3);
-		assertThat(testBean.calls.get(6).intValue()).isEqualTo(4);
-		assertThat(testBean.calls.get(7).intValue()).isEqualTo(4);
-		assertThat(testBean.calls.get(8).intValue()).isEqualTo(5);
-		assertThat(testBean.calls.get(9).intValue()).isEqualTo(5);
-		testBean.reset();
-		channel.send(new GenericMessage<String>("test-11"));
-		assertThat(testBean.calls.size()).isEqualTo(1);
-		assertThat(testBean.calls.get(0).intValue()).isEqualTo(1);
-		context.close();
+		try (GenericApplicationContext context = TestUtils.createTestApplicationContext()) {
+			context.registerBeanDefinition("postProcessor",
+					new RootBeanDefinition(MessagingAnnotationPostProcessor.class));
+			BeanDefinitionBuilder channelBuilder = BeanDefinitionBuilder.rootBeanDefinition(DirectChannel.class);
+			channelBuilder.addConstructorArgValue(null);
+			RootBeanDefinition channelDefinition = (RootBeanDefinition) channelBuilder.getBeanDefinition();
+			context.registerBeanDefinition("input", channelDefinition);
+			RootBeanDefinition testBeanDefinition = new RootBeanDefinition(TestBean.class);
+			testBeanDefinition.getConstructorArgumentValues().addGenericArgumentValue(2);
+			context.registerBeanDefinition("testBean", testBeanDefinition);
+			context.refresh();
+			TestBean testBean = (TestBean) context.getBean("testBean");
+			MessageChannel channel = (MessageChannel) context.getBean("input");
+			channel.send(new GenericMessage<>("test-1"));
+			channel.send(new GenericMessage<>("test-2"));
+			channel.send(new GenericMessage<>("test-3"));
+			channel.send(new GenericMessage<>("test-4"));
+			channel.send(new GenericMessage<>("test-5"));
+			channel.send(new GenericMessage<>("test-6"));
+			channel.send(new GenericMessage<>("test-7"));
+			channel.send(new GenericMessage<>("test-8"));
+			channel.send(new GenericMessage<>("test-9"));
+			channel.send(new GenericMessage<>("test-10"));
+			assertThat(testBean.calls.size()).isEqualTo(10);
+			assertThat(testBean.calls.get(0).intValue()).isEqualTo(1);
+			assertThat(testBean.calls.get(1).intValue()).isEqualTo(1);
+			assertThat(testBean.calls.get(2).intValue()).isEqualTo(2);
+			assertThat(testBean.calls.get(3).intValue()).isEqualTo(2);
+			assertThat(testBean.calls.get(4).intValue()).isEqualTo(3);
+			assertThat(testBean.calls.get(5).intValue()).isEqualTo(3);
+			assertThat(testBean.calls.get(6).intValue()).isEqualTo(4);
+			assertThat(testBean.calls.get(7).intValue()).isEqualTo(4);
+			assertThat(testBean.calls.get(8).intValue()).isEqualTo(5);
+			assertThat(testBean.calls.get(9).intValue()).isEqualTo(5);
+			testBean.reset();
+			channel.send(new GenericMessage<>("test-11"));
+			assertThat(testBean.calls.size()).isEqualTo(1);
+			assertThat(testBean.calls.get(0).intValue()).isEqualTo(1);
+		}
 	}
 
 	@Test
 	public void directChannelAndRoundRobinDispatcher() {
-		GenericApplicationContext context = TestUtils.createTestApplicationContext();
-		context.registerBeanDefinition("postProcessor", new RootBeanDefinition(MessagingAnnotationPostProcessor.class));
-		RootBeanDefinition channelDefinition = new RootBeanDefinition(DirectChannel.class);
-		channelDefinition.getConstructorArgumentValues().addGenericArgumentValue(new RoundRobinLoadBalancingStrategy());
-		context.registerBeanDefinition("input", channelDefinition);
-		RootBeanDefinition testBeanDefinition = new RootBeanDefinition(TestBean.class);
-		testBeanDefinition.getConstructorArgumentValues().addGenericArgumentValue(1000);
-		context.registerBeanDefinition("testBean", testBeanDefinition);
-		context.refresh();
-		TestBean testBean = (TestBean) context.getBean("testBean");
-		MessageChannel channel = (MessageChannel) context.getBean("input");
-		channel.send(new GenericMessage<String>("test-1"));
-		channel.send(new GenericMessage<String>("test-2"));
-		channel.send(new GenericMessage<String>("test-3"));
-		channel.send(new GenericMessage<String>("test-4"));
-		channel.send(new GenericMessage<String>("test-5"));
-		List<Integer> calls = testBean.calls;
-		assertThat(calls.size()).isEqualTo(5);
-		assertThat(calls.get(0).intValue()).isEqualTo(1);
-		assertThat(calls.get(1).intValue()).isEqualTo(2);
-		assertThat(calls.get(2).intValue()).isEqualTo(3);
-		assertThat(calls.get(3).intValue()).isEqualTo(4);
-		assertThat(calls.get(4).intValue()).isEqualTo(5);
-		context.close();
+		try (GenericApplicationContext context = TestUtils.createTestApplicationContext()) {
+			context.registerBeanDefinition("postProcessor",
+					new RootBeanDefinition(MessagingAnnotationPostProcessor.class));
+			RootBeanDefinition channelDefinition = new RootBeanDefinition(DirectChannel.class);
+			channelDefinition.getConstructorArgumentValues()
+					.addGenericArgumentValue(new RoundRobinLoadBalancingStrategy());
+			context.registerBeanDefinition("input", channelDefinition);
+			RootBeanDefinition testBeanDefinition = new RootBeanDefinition(TestBean.class);
+			testBeanDefinition.getConstructorArgumentValues().addGenericArgumentValue(1000);
+			context.registerBeanDefinition("testBean", testBeanDefinition);
+			context.refresh();
+			TestBean testBean = (TestBean) context.getBean("testBean");
+			MessageChannel channel = (MessageChannel) context.getBean("input");
+			channel.send(new GenericMessage<>("test-1"));
+			channel.send(new GenericMessage<>("test-2"));
+			channel.send(new GenericMessage<>("test-3"));
+			channel.send(new GenericMessage<>("test-4"));
+			channel.send(new GenericMessage<>("test-5"));
+			List<Integer> calls = testBean.calls;
+			assertThat(calls.size()).isEqualTo(5);
+			assertThat(calls.get(0).intValue()).isEqualTo(1);
+			assertThat(calls.get(1).intValue()).isEqualTo(2);
+			assertThat(calls.get(2).intValue()).isEqualTo(3);
+			assertThat(calls.get(3).intValue()).isEqualTo(4);
+			assertThat(calls.get(4).intValue()).isEqualTo(5);
+		}
 	}
